+++ config/scribe.php
@@ -241,7 +241,13 @@ INTRO
             Strategies\Responses\UseResponseFileTag::class,
             [
                 Strategies\Responses\ResponseCalls::class,
-                ['only' => ['GET *']]
+                [
+                    'only' => ['GET *'],
+                    // Disable debug mode when generating response calls to avoid error stack traces in responses
+                    'config' => [
+                        'app.debug' => false,
+                    ],
+                ]
             ]
         ],
         'responseFields' => [