+++ src/Tools/Utils.php
@@ -9,6 +9,7 @@ use FastRoute\RouteParser\Std;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Routing\Route;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
 use Knuckles\Scribe\Exceptions\CouldntFindFactory;
 use Knuckles\Scribe\Exceptions\CouldntGetRouteDetails;
@@ -191,6 +192,11 @@ class Utils
         }
     }
 
+    public static function makeDirectoryRecursive(string $dir): void
+    {
+        File::isDirectory($dir) || File::makeDirectory($dir, 0777, true, true);
+    }
+
     public static function deleteFilesMatching(string $dir, callable $condition): void
     {
         if (class_exists(LocalFilesystemAdapter::class)) {
@@ -364,7 +370,7 @@ class Utils
     {
         // We only load our custom translation layer if we really need it
         if (!ScribeServiceProvider::$customTranslationLayerLoaded) {
-            (new ScribeServiceProvider(app()))->loadCustomTranslationLayer();
+            app(ScribeServiceProvider::class, ['app' => app()])->loadCustomTranslationLayer();
         }
 
         $translation = trans($key, $replace);